bump docker-compose.yml for the site to work with dev properly
Previously, an image built for the `build` service was then used
for `server`, and changes were invisible to the user
before the container rebuild.

After that change, the useless static `build` service is deleted,
Dockerfile is only used in the CI pipeline, and only the `server`
service is left in docker-compose for the user to test
and see documentation changes locally in real-time.

Resolves #1178